Move attachment datagen to its own package
